move block filter predicate to outside function
